+++ transformers/llm/engine/app/mls_config.hpp
@@ -1,0 +1,10 @@
+//
+// Created by ruoyi.sjd on 2024/12/25.
+// Copyright (c) 2024 Alibaba Group Holding Limited All rights reserved.
+//
+
+#pragma once
+
+namespace mls {
+   const char* const kCachePath = "~/.mnnmodels";
+}
